pyproject.toml
[project]
name = "quiz-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-wtf>=1.2.2",
    "litellm>=1.67.1",
    "python-dotenv>=1.1.0",
]
